TCIC2-11 Moved some things, most notably the oeis lookup was moved to a proper API module
