+++ chatbot/src/responder/core/info.rs
