+++ chatbot/src/responder/core/niceties.rs
